TDHeat05 {statgenSTA}	R Documentation

Field data for a wheat experiment in Mexico

Description

A dataset converted to a TD object containing raw plot data for one trial from a series of wheat trials conducted in Mexico by CIMMYT. The different trials took place under different regimes of irrigation and temperature, there were 4 trials across two years, labelled as DRIP05, HEAT05, HEAT06 and IRRI06. The TD object only contains the data for HEAT05. Within each trial, a set of 167 progeny of a RIL (Recombinant Inbred Line; 8 generations) population were tested alongside the population parents (Seri and Babax). A lattice design with two replicates was used for each trial. In the first replicate the entries were not randomized, as they were considered to be a random selection from a population.

Usage

TDHeat05

Format

A TD object, a list containing a data.frame with the following columns:

trial

trial, a combination of watering regime, year and nitrogen treatment

genotype

genotype

Plot

plot number in the field

repId

replicate

subBlock

block id

rowId

row within the field (as factor)

colId

column within the field (as factor)

yield

yield in grams per square meter

rowCoord

row within the field (as numerical value)

colCoord

column within the field (as numerical value)

[Package statgenSTA version 1.0.15 Index]