summary.STA {statgenSTA}	R Documentation

Summarizing objects of class STA

Description

summary method for class STA.

Usage

## S3 method for class 'STA'
summary(
  object,
  trials = NULL,
  trait = NULL,
  nBest = 20,
  sortBy = NULL,
  naLast = TRUE,
  decreasing = TRUE,
  ...
)

Arguments

object
An object of class STA.

trials
A character vector indicating the trial to summarize. If trial = NULL a summary is made of all trials in the STA object. If a single trial is selected a full summary for this trial is created. For multiple trials a summary table with the most importantant statistics is returned.

trait
A character string indicating the trait to summarize. If trait = NULL and only one trait is modeled, this trait is summarized.

nBest
An integer indicating the number of the best genotypes (sorted by either BLUEs or BLUPs) to print. If NA, all genotypes will be printed.

sortBy
A character string specifying how the genotypes will be sorted. Either "BLUEs", "BLUPs" or NA (i.e. no sorting).

naLast
Should missing values in the data be put last when sorting?

decreasing
Should the sort order be decreasing?

...
Further arguments - not used.

See Also

Other functions for STA objects: STAtoCross(), STAtoTD(), plot.STA(), report.STA()

Examples

## Run a single trait analysis using SpATS.
modSp <- fitTD(TD = TDHeat05,
               design = "res.rowcol",
               traits = "yield")

## Print a summary of the fitted model.
summary(modSp)

[Package statgenSTA version 1.0.15 Index]