Summarizing objects of class STA

summary method for class STA.

Usage

# S3 method for class 'STA'
summary(
  object,
  trials = NULL,
  trait = NULL,
  nBest = 20,
  sortBy = NULL,
  naLast = TRUE,
  decreasing = TRUE,
  ...
)

Arguments

object

An object of class STA.

trials

A character vector indicating the trial to summarize. If trial = NULL a summary is made of all trials in the STA object. If a single trial is selected a full summary for this trial is created. For multiple trials a summary table with the most importantant statistics is returned.

trait

A character string indicating the trait to summarize. If trait = NULL and only one trait is modeled, this trait is summarized.

nBest

An integer indicating the number of the best genotypes (sorted by either BLUEs or BLUPs) to print. If NA, all genotypes will be printed.

sortBy

A character string specifying how the genotypes will be sorted. Either "BLUEs", "BLUPs" or NA (i.e. no sorting).

naLast

Should missing values in the data be put last when sorting?

decreasing

Should the sort order be decreasing?

...

Further arguments - not used.

See also

Examples

## Run a single trait analysis using SpATS.
modSp <- fitTD(TD = TDHeat05,
               design = "res.rowcol",
               traits = "yield")
#> Using SpATS for fitting models.

## Print a summary of the fitted model.
summary(modSp)
#> Summary statistics 
#> ==================
#> 
#> Summary statistics for yield in HEAT05  
#> 
#>                             yield
#> Number of observations      338.0
#> Number of missing values      0.0
#> Mean                       184.16
#> Median                     179.81
#> Min                         54.58
#> Max                        341.77
#> First quantile             148.56
#> Third quantile             220.50
#> Variance                 2569.229
#> 
#> 
#> Estimated heritability 
#> ======================
#> 
#> Heritability: 0.92 
#> 
#> Predicted means (BLUEs & BLUPs) 
#> ===============================
#> Best 20 genotypes
#>         BLUEs      SE   BLUPs     SE
#> SB023 285.017  11.120 273.734 10.668
#> SB025 283.928  11.058 273.581 10.586
#> SB022 281.899  11.149 269.827 10.701
#> SB032 278.929  11.100 271.785 10.624
#> SB011 271.102  11.300 264.319 10.780
#> SB156 267.220  11.313 257.920 10.809
#> SB144 265.796  11.391 257.498 10.862
#> SB031 252.946  11.063 247.454 10.581
#> SB093 247.711  11.249 245.990 10.761
#> SB054 247.144  11.463 239.404 10.948
#> SB055 247.041  11.414 240.557 10.903
#> SB020 241.192  11.094 233.505 10.630
#> SB028 239.224  11.212 234.370 10.720
#> SB128 238.020  11.437 232.843 10.899
#> SB194 237.852  11.260 232.810 10.765
#> SB155 236.512  11.268 229.693 10.765
#> SB143 236.263  11.454 229.600 10.917
#> SB027 235.379  11.120 229.517 10.607
#> SB163 234.342  11.146 229.361 10.658
#> SB181 233.111  11.083 229.460 10.588