Summarizing objects of class STA

summary method for class STA.

# S3 method for STA
summary(
  object,
  trials = NULL,
  trait = NULL,
  nBest = 20,
  sortBy = NULL,
  naLast = TRUE,
  decreasing = TRUE,
  ...
)

Arguments

object
An object of class STA.

trials
A character vector indicating the trial to summarize. If trial = NULL a summary is made of all trials in the STA object. If a single trial is selected a full summary for this trial is created. For multiple trials a summary table with the most importantant statistics is returned.

trait
A character string indicating the trait to summarize. If trait = NULL and only one trait is modeled, this trait is summarized.

nBest
An integer indicating the number of the best genotypes (sorted by either BLUEs or BLUPs) to print. If NA, all genotypes will be printed.

sortBy
A character string specifying how the genotypes will be sorted. Either "BLUEs", "BLUPs" or NA (i.e. no sorting).

naLast
Should missing values in the data be put last when sorting?

decreasing
Should the sort order be decreasing?

...
Further arguments - not used.

See also

Examples

## Run a single trait analysis using SpATS.
myModel <- fitTD(TD = TDHeat05, design = "res.rowcol", traits = "yield")

#> Using SpATS for fitting models.

## Print a summary of the fitted model.
summary(myModel)

#> Summary statistics 
#> ==================
#> 
#> Summary statistics for yield in HEAT05  
#> 
#>                             yield
#> Number of observations        338
#> Number of missing values        0
#> Mean                       184.16
#> Median                     179.81
#> Min                         54.58
#> Max                        341.77
#> First quantile             148.56
#> Third quantile             220.50
#> Variance                 2569.229
#> 
#> 
#> Estimated heritability 
#> ======================
#> 
#> Heritability: 0.92 
#> 
#> Predicted means (BLUEs & BLUPs) 
#> ===============================
#> Best 20 genotypes
#>         BLUEs      SE   BLUPs     SE
#> SB023 285.060  11.181 273.220 10.717
#> SB025 284.271  11.079 273.493 10.600
#> SB022 281.860  11.241 269.082 10.787
#> SB032 278.754  11.111 271.306 10.630
#> SB011 270.475  11.319 263.558 10.813
#> SB156 266.811  11.300 257.270 10.776
#> SB144 265.166  11.354 256.617 10.813
#> SB031 253.335  11.056 247.661 10.567
#> SB093 247.562  11.197 246.404 10.729
#> SB055 246.730  11.390 239.976 10.864
#> SB054 246.577  11.474 238.363 10.930
#> SB020 241.319  11.159 232.928 10.684
#> SB028 239.307  11.241 235.167 10.727
#> SB194 238.182  11.300 232.570 10.836
#> SB128 237.677  11.343 232.245 10.831
#> SB155 236.158  11.255 229.047 10.723
#> SB027 235.905  11.125 230.022 10.604
#> SB143 235.825  11.390 229.027 10.848
#> SB163 233.971  11.130 228.848 10.628
#> SB181 233.137  11.084 229.614 10.580